I added a model response validation and I create a view to handle the api call
