+++ app/endpoints/weather_handler.py
@@ -1,13 +1,19 @@
-from fastapi import APIRouter
-from app.models.weather import Weather
+from fastapi import APIRouter, Depends
+from app.models.weather import Location,WeatherResponse
+from .views.weather_view import Weather
 
 router = APIRouter()
 
-@router.get('/')
-async def Wheater_by_city_state(
-        city:str, country:str
-    ) -> Weather:
-        weahter_instance = Weather(city=city, country=country)
+@router.get('/', response_model=WeatherResponse)
+async def Wheater_by_city_state(location: Location = Depends()):
+    result = ''
+    try:
+        weahter_instance = Weather(
+            city=location.city,
+            country=location.country
+        )
         result = weahter_instance.get_from_open_weather()
-        
-        return result
+    except Exception as e:
+        print(e)
+
+    return result
